+++ opentelemetry-api/src/metrics/instruments/histogram.rs
@@ -1,6 +1,6 @@
 use crate::{
     metrics::{InstrumentBuilder, MetricsError},
-    Context, KeyValue,
+    KeyValue,
 };
 use core::fmt;
 use std::convert::TryFrom;
@@ -9,7 +9,7 @@ use std::sync::Arc;
 /// An SDK implemented instrument that records a distribution of values.
 pub trait SyncHistogram<T> {
     /// Adds an additional value to the distribution.
-    fn record(&self, cx: &Context, value: T, attributes: &[KeyValue]);
+    fn record(&self, value: T, attributes: &[KeyValue]);
 }
 
 /// An instrument that records a distribution of values.
@@ -32,8 +32,8 @@ impl<T> Histogram<T> {
     }
 
     /// Adds an additional value to the distribution.
-    pub fn record(&self, cx: &Context, value: T, attributes: &[KeyValue]) {
-        self.0.record(cx, value, attributes)
+    pub fn record(&self, value: T, attributes: &[KeyValue]) {
+        self.0.record(value, attributes)
     }
 }