+++ opentelemetry-sdk/src/testing/metrics/in_memory_exporter.rs
@@ -47,7 +47,7 @@ use std::sync::{Arc, Mutex};
 ///  let meter = meter_provider.meter(std::borrow::Cow::Borrowed("example"));
 ///  let cx = Context::new();
 ///  let counter = meter.u64_counter("my_counter").init();
-///  counter.add(&cx, 1, &[KeyValue::new("key", "value")]);
+///  counter.add(1, &[KeyValue::new("key", "value")]);
 ///
 ///  meter_provider.force_flush(&cx).unwrap();
 ///