+++ opentelemetry/src/lib.rs
@@ -89,9 +89,7 @@
 //! ```
 //! # #[cfg(feature = "metrics")]
 //! # {
-//! use opentelemetry::{global, Context, KeyValue};
-//!
-//! let cx = Context::current();
+//! use opentelemetry::{global, KeyValue};
 //!
 //! // get a meter from a provider
 //! let meter = global::meter("my_service");
@@ -100,7 +98,7 @@
 //! let counter = meter.u64_counter("my_counter").init();
 //!
 //! // record a measurement
-//! counter.add(&cx, 1, &[KeyValue::new("http.client_ip", "83.164.160.102")]);
+//! counter.add(1, &[KeyValue::new("http.client_ip", "83.164.160.102")]);
 //! # }
 //! ```
 //!