+++ stress/src/metrics.rs
@@ -1,15 +1,11 @@
 use lazy_static::lazy_static;
-use opentelemetry_api::{
-    metrics::{Counter, MeterProvider as _},
-    Context,
-};
+use opentelemetry_api::metrics::{Counter, MeterProvider as _};
 use opentelemetry_sdk::metrics::{ManualReader, MeterProvider};
 use std::borrow::Cow;
 
 mod throughput;
 
 lazy_static! {
-    static ref CONTEXT: Context = Context::new();
     static ref PROVIDER: MeterProvider = MeterProvider::builder()
         .with_reader(ManualReader::builder().build())
         .build();
@@ -24,5 +20,5 @@ fn main() {
 }
 
 fn test_counter() {
-    COUNTER.add(&CONTEXT, 1, &[]);
+    COUNTER.add(1, &[]);
 }